ops: delete legacy ops package
The `ops-bedrock` package will likely be migrated to
the `ops` package. Keep around the dockerfile because the
builds of the `fault-detector` and the `chain-mon` depend
on this dockerfile. We should eventually refactor away this
dockerfile and follow the pattern that the other packages use
where the dockerfile is in the particular package instead of
being outside in its own package.