cleanup: delete custom ethers fallback provider
Deletes a no longer used fallback provider that was
mostly copy pasted from ethers v5. There was no good
way to specify the headers and we needed to do so,
so we used this client to ensure that headers can
be passed. It is no longer used so we can safely delete
it.