contracts-bedrock: tightly pack creation timestamp (#6020)
Pack the creation timestamp in the storage slot along
with the address of the dispute games so that the
information can be accessed easily by services that
query the contract.

A service looking for recent dispute games can
binary search or linear search backwards using the
`gameAtIndex` function. It is guaranteed that the
timestamp will always increase or be equal as the
index increases.