contracts-bedrock: optimize storage layout checks
Optimize the storage layout checks by removing a dep
on hardhat. Instead of reading massive build-info files
from disk, simply read the forge artifacts directly and
then filter through the interesting ones.
This now only takes a few seconds to run when the compilation
is cached.

Also delete a dead hh task for doing deposits, this was built
as a PoC to trigger a deposit.