contracts-bedrock: delete dead script
This script is no longer used and generates no artifacts. The artifacts
are currently exported by path and the JSON files can be imported.
While this style of artifact getter API is nice, it causes the bundle
size for frontends to be very large and they cannot shake all of the
unused artifacts out since any can be requested at runtime.