op-challenger: Add tests for CLI args
Refactors config to make it testable and provide programmatic access to the default config.
Unused RPCConfig is removed.
Log config is no longer part of application config - a configured logger is supplied to the application.
Unused Metrics and PProf config is removed - these are not application concepts but rather a service to be passed in like logging.