contracts-bedrock: fix system config initialize
Every sentinel contract should have null values in their
storage to make them unusable. Sometimes there are constraints
on what these values can be in the initializer and they cannot
be exactly 0 but must be close enough to make the contract
useless. This commit updates the startBlock number so that
it is not 0 because if it is 0 then it will be set to `block.number`
which is not what we want.