op-node: Metrics in the channel bank
This adds 3 new metrics to channel bank:
1. New Head Channel Event
2. Channel Timed Out Event
3. Frame Added Event

The new channel head event is the most complex. It indicates that there is new
channel in the ready position of the channel bank. It does not indicate if the
channel is ready to be read from. It is complicated to implement because we
have several ways of removing channels which would mean that a new channel becomes
the head channel.

The channel timeout & frame added events are much simpler.

The intention is to make it easier to determine if the head channel in the channel
bank is not progressing.