deps: use node lts
Upgrade from node.js v16 to v18. v16 starts end of life on
9/11/2023, meaning it will no longer get security updates.
v18 is currently LTS, https://github.com/nodejs/release#release-schedule
for additional information.

We previously were not able to upgrade node.js versions because
we were using an older version of hardhat that did not support newer
versions of node.js but that is no longer the case.