Support old Make version by avoiding SHELLSTATUS
SHELLSTATUS is only available on GNU Make 4.2 (released on 2016-05-22)
and later, which is unfortunatly not installed on MacOS by default.

While I am highly annoyed that Apple is unwilling to update their Make
version and don't want encourage such behaviour, the change in this
commit avoids this requirement and seems even simpler than the original
code.