op-chain-ops: add additional assertions the `check-l2`
The `check-l2` script works against new networks but not legacy
networks. We need a solution that can work against both new networks
and legacy networks. This adds test coverage for new networks
as `check-l2` runs in CI against the local devnet.

Add test coverage for reinitialize on L2 contracts.
Also ensures that the `_initialized` value is set as
expected.