contracts-bedrock: pseudorandom blocknumber in portal fuzz test
Set a pseudorandom `block.number` when fuzzing the `OptimismPortal`
deposit transaction codepath. This is a small fix required for
the migration to using the deploy script to set up the unit tests.
This prevents underflows with computing the `_prevBlockNum` and also
adds additional converage since previously the block number was not
being set dynamically.

Breaking this out into its own smaller PR so that the deploy script
migration can be smaller.