contracts-bedrock: modern import style in tests
Migrate to the modern import style in the tests where the
path relative from the project root is used. This makes the
code much more portable and easier to modify without the dot
dot syntax. There should be very little old relative style
imports left in the codebase if any. There are no functional
changes with this commit.