ci-builder: go version env var
Makes setting the go version via the env var `GO_VERSION` possible.
Previously it hardcoded the version that it was unpacking but used
the env var in other locations. Now setting the env var for `GO_VERSION`
will allow the user to configure whatever version they want to include
in the image.