contracts-bedrock: simplify merkle trie setup
The `MerkleTrie` tests were using `CommonTest` but do not
require anything besides `FFIInterface` so instead of pulling
in all of the overhead of `CommonTest`, just deploy `FFIInterface`
on its on in the `setUp` function. This should be a slight
optimization as it reduces the amount of execution required for
setup.

Reducing diff from https://github.com/ethereum-optimism/optimism/pull/7928