contracts-bedrock: visibility into flake
A flake in CI was introduced when https://github.com/ethereum-optimism/optimism/pull/7928
was merged. It is probably due to a race condition when reading a file
from disk. Is there a way to have foundry only do something once for the
entire test suite? Tried moving things to the constructor instead of
`setUp` but that did not work. Ideally we do not need to read the file
from disk for each contract deployed, this adds a lot of overhead.
A solution around this is to refactor the way that the deploy script
works or to add in the env var that will skip the check that sometimes
fails.