contracts-bedrock: update `StorageSetter`
Adds a new setter that takes a list of slots to set. This allows a
single `upgradeToAndCall` to be used with this contract that sets many
storage slots in a single call. This will simplify future upgrades
using the 2 step upgrade model.

Includes fuzz tests for the new functionality. The complexity with
the fuzz testing of this operation is that we cannot pass in duplicate
keys, otherwise it is impossible to make assertions against because
there is no way to hook in to the call. Perhaps if we had a way to see
every state diff then we would be able to make assertions against that.