monorepo: submodules makefile
Update the `submodules` recipe in the makefile.
Trying to remove ignoring the command when in CI
because it looks for a github actions env var but
we now run in circleci so its possible that its
tech debt. Will add that check back in if it breaks
ci.

Fixes https://github.com/ethereum-optimism/optimism/issues/8132