fix(sdk): bug in supportsTokenPair
Fixes a bug in the supportsTokenPair where unsupported token interfaces
would cause the SDK to error out. Happens because the adapters check for
things that are only present in certain types of tokens. Adds a new
error handling check that wraps all sub-calls to supportsTokenPair in
each of the adapters so that the adapters don't need to worry about
this.