contracts-bedrock: delete state diff json
Removes the `deploy.json` state diff artifact. We don't necessarily
need to commit the state diff of the deploy script to the repo. We
should utilize the state diff of the deploy script when doing major
refactors of the deploy script to ensure that no bugs are being
introduced.

Adding in CI checks that the artifact has been updated is too much
overhead and not really worth the benefit. Certain changes to the deploy
script would result in huge diffs to the artifact without it being
particularly legible. This tooling is more useful for `superchain-ops`
where we should have every single storage slot diff fully accounted for.