op-node: fix ecotone upgrade test (#10092)
* op-node: fix ecotone upgrade test

Previously the upgrade transactions test made an assertion against
the bytecode that is in the bindings. This needs to be modified
because the upgrade txs cannot ever change as they are part of
consensus. Any modification to the bindings means that the assertion
will fail. We should test against the expected value directly
instead of going through the binding that has no guarantee
to be on a particular version and will change over time.

* op-node: fix test