op-e2e: use deposit method on weth (#10090)
Move away from using the fallback function on weth9 towards instead
using the explicit deposit method. This is to reduce diff with the
custom gas token project, just a very small change.

The weth98 contract doesn't have a fallback function, if it did
then this change would not be required.