op-bindings: remove (#10576)
* op-bindings: remove

Removes the `op-bindings` package from the monorepo. It is now
the responsibility of services to ensure that they are compatible
with particular versions of the contracts. This will reduce the
need for feature branches, as there doesn't have to be a single
implementation of a contract with its bindings.

The following link shows an easy way to generate bindings for
your own service: https://github.com/ethereum-optimism/optimism/blob/c26f841d6c95b280483a59391809b0fcda774f44/op-chain-ops/justfile

`abigen` from go-ethereum is very useful here.

* indexer: remove op-bindings from dockerfile

* op-e2e: cleanup

Cannot completely remove `legacybindings` because of
dependencies in types in different libraries.
Would be solved by a `go-sdk`.

* Revert "op-e2e: cleanup"

This reverts commit 43791f95213ad9566adeba087a022a48b794dc45.